Move code from BoogieGenerator.TrStatement into separate files
